18: Move TrackedGoalType to a types folder
