+++ react/src/types/GoalTypes.tsx
@@ -1,0 +1,11 @@
+export type TrackedGoalType = {
+  created_at: string
+  end_date: string | null
+  goal: string
+  goal_id: number
+  id: number
+  progress_rating: number | null
+  start_date: string | null
+  timeframe_type: string
+  updated_at: string
+}
